refactor: split main.py god-file into domain modules
1343-statement main.py split into focused modules with no circular imports:

  state.py      — shared mutable dicts, locks, load/save state
  config.py     — settings, keyring, telemetry, crash reports, tier model map
  schemas.py    — all Pydantic request/response models
  workspace.py  — _workspace(), _push_to_kernel(), path guards
  execution.py  — _execute_run, _execute_pipeline_step, _advance_pipeline,
                  _emit_pipeline, _resolve_llm, helpers
  routers/      — 9 FastAPI APIRouter modules (health, settings, ollama,
                  system, runs, pipelines, kernel, projects, uninstall)
  main.py       — app wiring only (~90 lines): FastAPI + CORS + middleware
                  + include_router × 9 + static mount

Dependency order (no cycles): stdlib → state → config → schemas →
workspace → execution → routers/* → main.

Tests updated: patch targets moved from agentsuitelocal.api.main.*
to their new homes (state._RUNS_FILE, state._runs, workspace._workspace,
execution._sanitize_qa_dimensions). All 120 backend tests pass, 64% coverage.
