+++ tests/test_integration.py
@@ -17,7 +17,7 @@ import time
 import httpx
 import pytest
 
-from agentsuitelocal.api.main import _runs
+from agentsuitelocal.api.state import _runs
 
 # live_server is session-scoped and defined in tests/conftest.py
